+++ src/functionaltests/java/com/ericsson/ei/notifications/ttl/TestTTLSteps.java
@@ -27,7 +27,7 @@ import org.springframework.util.SocketUtils;
 
 import com.ericsson.ei.exception.AuthenticationException;
 import com.ericsson.ei.handlers.MongoDBHandler;
-import com.ericsson.ei.subscription.InformSubscriber;
+import com.ericsson.ei.notifications.InformSubscriber;
 import com.ericsson.ei.utils.FunctionalTestBase;
 import com.ericsson.ei.utils.HttpRequest;
 import com.ericsson.ei.utils.HttpRequest.HttpMethod;
