+++ src/main/java/com/ericsson/ei/controller/model/ParseInstanceInfoEI.java
@@ -19,8 +19,8 @@ package com.ericsson.ei.controller.model;
 import com.ericsson.ei.erqueryservice.ERQueryService;
 import com.ericsson.ei.handlers.ObjectHandler;
 import com.ericsson.ei.handlers.RmqHandler;
-import com.ericsson.ei.subscription.InformSubscriber;
-import com.ericsson.ei.subscription.EmailSender;
+import com.ericsson.ei.notifications.EmailSender;
+import com.ericsson.ei.notifications.InformSubscriber;
 import com.ericsson.ei.subscription.SubscriptionHandler;
 import com.ericsson.ei.waitlist.WaitListStorageHandler;
 import lombok.Getter;
