+++ src/main/java/com/ericsson/ei/notifications/EmailSender.java
@@ -14,7 +14,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package com.ericsson.ei.subscription;
+package com.ericsson.ei.notifications;
 
 import java.util.HashSet;
 import java.util.Set;
