+++ src/test/java/com/ericsson/ei/subscription/SubscriptionHandlerTest.java
@@ -55,6 +55,9 @@ import org.springframework.util.MultiValueMap;
 import com.ericsson.ei.App;
 import com.ericsson.ei.controller.model.QueryResponse;
 import com.ericsson.ei.jmespath.JmesPathInterface;
+import com.ericsson.ei.notifications.EmailSender;
+import com.ericsson.ei.notifications.HttpRequestSender;
+import com.ericsson.ei.notifications.InformSubscriber;
 import com.ericsson.ei.handlers.MongoDBHandler;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
